Cleanup and refactoring; update search query based on search input
