+++ apps/kg-explorer/src/app/pages/main-page/main-page.component.html
@@ -22,7 +22,7 @@
       [filterCategories]="filterCategories()"
       (toggleForm)="sidenav.toggle()"
       [formClosed]="!sidenav.opened"
-      (formChanges)="handleFilterChanges($event)"
+      (formChanges)="handleFilterSelectionChanges($event)"
       [currentFilters]="filters()"
     ></hra-filter-menu>
   </mat-sidenav>
